@@ -4,6 +4,7 @@ import fr.theoszanto.mc.express.ExpressObject;
 import fr.theoszanto.mc.express.ExpressPlugin;
 import fr.theoszanto.mc.express.utils.ItemBuilder;
 import fr.theoszanto.mc.express.utils.MathUtils;
+import net.kyori.adventure.text.Component;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
@@ -27,7 +28,7 @@ public abstract class ExpressGUI<P extends ExpressPlugin<P>> extends ExpressObje
 
 	public ExpressGUI(@NotNull P plugin, int rows, @NotNull String key, @Nullable Object @NotNull... format) {
 		this(plugin, Bukkit.createInventory(null, MathUtils.minMax(1, rows, 6) * 9,
-				plugin.i18n("menu.title-style", "title", plugin.i18n(key, format))), MathUtils.minMax(1, rows, 6));
+				Component.text(plugin.i18n("menu.title-style", "title", plugin.i18n(key, format)))), MathUtils.minMax(1, rows, 6));
 	}
 
 	private ExpressGUI(@NotNull P plugin, @NotNull Inventory inventory, int rows) {
@@ -122,7 +123,7 @@ public abstract class ExpressGUI<P extends ExpressPlugin<P>> extends ExpressObje
 
 	protected static boolean impactOtherInventory(@NotNull InventoryAction action) {
 		return action == InventoryAction.MOVE_TO_OTHER_INVENTORY || action == InventoryAction.COLLECT_TO_CURSOR
-				|| action == InventoryAction.HOTBAR_MOVE_AND_READD || action == InventoryAction.UNKNOWN;
+				|| action == InventoryAction.HOTBAR_SWAP || action == InventoryAction.UNKNOWN;
 	}
 
 	protected static int slot(int row, int column) {
